+++ components/src/NavBar/Branding.js
